refactor: rename helpers as reference_implementations
