+++ tests/kat_tests/external_estimators/reference_implementations/MQEstimator/mpkc/series/nmonomial.py
