+++ tests/kat_tests/external_estimators/reference_implementations/attack_cost.sage
