+++ tests/kat_tests/external_estimators/reference_implementations/cost.sage
