+++ tests/kat_tests/kat.yaml
@@ -284,7 +284,7 @@ le:
       - 61.142121724263994
     - !!python/tuple
       - !!python/tuple [100, 52, 7]
-      - 56.9431599667116
+      - 55.85294020421421
     - !!python/tuple
       - !!python/tuple [100, 52, 11]
       - 61.20412589171779
